feat: move Button component to globalStyle
