+++ src/styles/Global.js
@@ -1,4 +1,5 @@
-import { createGlobalStyle } from 'styled-components';
+import { Link } from 'react-router-dom';
+import styled, { createGlobalStyle } from 'styled-components';
 import { normalize } from 'styled-normalize';
 
 const GlobalStyles = createGlobalStyle`
@@ -15,11 +16,11 @@ const GlobalStyles = createGlobalStyle`
   }
   h1{
     line-height: ${({ theme }) => theme.fonts.lineHeight.h1};
-     font-size: ${({ theme }) => theme.fonts.size.h1};
+    font-size: ${({ theme }) => theme.fonts.size.h1};
   }
   h2{
-      line-height: ${({ theme }) => theme.fonts.lineHeight.h2};
-       font-size: ${({ theme }) => theme.fonts.size.h2};
+    line-height: ${({ theme }) => theme.fonts.lineHeight.h2};
+    font-size: ${({ theme }) => theme.fonts.size.h2};
   }
   p{
     margin:0;
@@ -28,6 +29,20 @@ const GlobalStyles = createGlobalStyle`
   }
   a{
     text-decoration: none;
+  }
+`;
+export const Button = styled(Link)`
+  font-size: ${({ theme }) => theme.fonts.size.smaller};
+  line-height: ${({ theme }) => theme.fonts.lineHeight.shorter};
+  background-color: ${({ theme }) => theme.colors.accent};
+  border: .1px ${({ theme }) => theme.colors.accent} solid; 
+  border-radius: 4px;
+  color:  ${({ theme }) => theme.colors.white};
+  padding: 0.844rem 0.927rem 0.844rem 0.948rem;
+  margin-bottom: 2.875rem;
+  &:hover{
+          background-color: white;
+          color:${({ theme }) => theme.colors.accent};
   }
 `;
